+++ tests/scar_add_metadata_toolbox_tests/classes/__init__.py
@@ -1,10 +1,10 @@
 from __future__ import annotations
 
 from pathlib import Path
-from typing import Any
 
 from bas_metadata_library.standards.iso_19115_2 import MetadataRecord
-from flask import Flask, Request, Response
+from flask import Request, Response
+from flask_entra_auth.token import EntraToken
 
 from scar_add_metadata_toolbox.csw import (
     CSWAmbiguousRequestError,
@@ -26,7 +26,6 @@ from scar_add_metadata_toolbox.csw import (
     RecordNotFoundError,
     RecordServerError,
 )
-from scar_add_metadata_toolbox.placeholders import PlaceholderAzureToken
 
 
 class MockCSWClient(CSWClient):
@@ -182,17 +181,17 @@ class MockCSWServer(CSWServer):
             raise CSWTrackingRepositoryAlreadyInitialisedError() from None
         self.backing_repo_initialised = True
 
-    def process_request(self, request: Request, token: PlaceholderAzureToken | None = None) -> Response:
+    def process_request(self, request: Request, token: EntraToken | None = None) -> Response:
         return Response("ok")
 
 
 class MockCSWServerBackingDBNotSetup(MockCSWServer):
-    def process_request(self, request: Request, token: PlaceholderAzureToken | None = None) -> Response:
+    def process_request(self, request: Request, token: EntraToken | None = None) -> Response:
         raise CSWDatabaseNotInitialisedError() from None
 
 
 class MockCSWServerBackingRepoNotSetup(MockCSWServer):
-    def process_request(self, request: Request, token: PlaceholderAzureToken | None = None) -> Response:
+    def process_request(self, request: Request, token: EntraToken | None = None) -> Response:
         raise CSWTrackingRepositoryNotInitialisedError() from None
 
 
@@ -207,79 +206,35 @@ class MockCSWServerRevisionTrackingInvalidCredentials(MockCSWServer):
 
 
 class MockCSWServerNoRequestType(MockCSWServer):
-    def process_request(self, request: Request, token: PlaceholderAzureToken | None = None) -> Response:
+    def process_request(self, request: Request, token: EntraToken | None = None) -> Response:
         raise CSWUnknownRequestError() from None
 
 
 class MockCSWServerAmbiguousRequestError(MockCSWServer):
-    def process_request(self, request: Request, token: PlaceholderAzureToken | None = None) -> Response:
+    def process_request(self, request: Request, token: EntraToken | None = None) -> Response:
         raise CSWAmbiguousRequestError() from None
 
 
 class MockCSWServerUnmappedRequestError(MockCSWServer):
-    def process_request(self, request: Request, token: PlaceholderAzureToken | None = None) -> Response:
+    def process_request(self, request: Request, token: EntraToken | None = None) -> Response:
         raise CSWUnmappedRequestError() from None
 
 
 class MockCSWServerAuthTokenError(MockCSWServer):
-    def process_request(self, request: Request, token: PlaceholderAzureToken | None = None) -> Response:
+    def process_request(self, request: Request, token: EntraToken | None = None) -> Response:
         raise CSWAuthError() from None
 
 
 class MockCSWServerMissingAuthToken(MockCSWServer):
-    def process_request(self, request: Request, token: PlaceholderAzureToken | None = None) -> Response:
+    def process_request(self, request: Request, token: EntraToken | None = None) -> Response:
         raise CSWAuthMissingError() from None
 
 
 class MockCSWServerInsufficientAuthToken(MockCSWServer):
-    def process_request(self, request: Request, token: PlaceholderAzureToken | None = None) -> Response:
+    def process_request(self, request: Request, token: EntraToken | None = None) -> Response:
         raise CSWAuthInsufficientError() from None
 
 
 class MockCSWServerRequestsFail(MockCSWServer):
-    def process_request(self, request: Request, token: PlaceholderAzureToken | None = None) -> Response:
+    def process_request(self, request: Request, token: EntraToken | None = None) -> Response:
         raise RecordServerError() from None
-
-
-class MockPublicClientApplication:
-    # noinspection PyUnusedLocal
-    def __init__(self, client_id: str, authority: str) -> None:
-        pass
-
-    # noinspection PyUnusedLocal
-    @staticmethod
-    def initiate_device_flow(scopes: list[str]) -> dict:
-        return {"user_code": "test"}
-
-    # noinspection PyUnusedLocal
-    @staticmethod
-    def acquire_token_by_device_flow(device_flow: dict) -> dict:
-        return {"access_token": None}
-        # return {
-        #     "access_token": TestJwt(
-        #         app=current_app, roles=["BAS.MAGIC.ADD.Records.ReadWrite.All", "BAS.MAGIC.ADD.Records.Publish.All"]  # noqa: ERA001
-        #     ).dumps()
-        # }  noqa: ERA001
-
-
-def create_mock_auth(scopes_used: list[str] | None = None):
-    class MockFlaskAzureOauth:
-        # noinspection PyUnusedLocal
-        def init_app(self, app: Flask) -> None:
-            pass
-
-        def __call__(self, scopes: str | None = None, *args, **kwargs) -> callable:  # noqa: ANN002, ANN003
-            def checkauth(func: callable) -> Any:
-                def wrapper(*args, **kwargs) -> callable:  # noqa: ANN002, ANN003
-                    if scopes_used is not None:
-                        scopes_used.append(scopes)
-
-                    return func(*args, **kwargs)
-
-                wrapper.__name__ = func.__name__
-
-                return wrapper
-
-            return checkauth
-
-    return MockFlaskAzureOauth